Implemented database.

+++ myrustchat/test.sh
@@ -1,7 +1,17 @@
 #!/bin/bash
-xterm -T server -e bash  -c 'cargo run --bin server; read' & > /dev/null 
+
+dir=$(dirname $0)
+server="cargo run --bin server --"
+client="cargo run --bin client --"
+
+if ! [ -f "$dir/server.db" ]; then
+    $server register -u Alice -p aaa || exit 1
+    $server register -u Bob -p bbb || exit 1
+fi
+
+xterm -T server -e bash  -c 'cargo run --bin server -- run; read' & > /dev/null 
 sleep 1
-xterm -T client1 -e bash  -c 'cargo run --bin client -- -n Alice; read' & > /dev/null
-xterm -T client2 -e bash  -c 'cargo run --bin client -- -n Bob; read' & > /dev/null
+xterm -T 'Client Alice' -e bash  -c 'cargo run --bin client -- -u Alice -p aaa; read' & > /dev/null
+xterm -T 'Client Bob' -e bash  -c 'cargo run --bin client -- -u Bob -p bbb; read' & > /dev/null
 trap "trap - SIGTERM && kill -- -$$" SIGINT SIGTERM EXIT
 sleep 99999
